awx/plugins/collections/ansible_collections/openstack/cloud/test-requirements.txt
pbr
ansible
pycodestyle
flake8
pylint
voluptuous
yamllint
rstcheck
ruamel.yaml
